# api/api.yaml
openapi: 3.0.3
info:
  title: VCBackend
  description: Backend component to verify credentials
  version: 0.0.1
...
paths:
  /issuer/api/v1/newcredential:
    get:
      tags:
        - issuer
        - frontend
      operationId: IssuerPageNewCredentialFormDisplay
      summary: Returns the page to issue new credentials
      responses:
        '200':
          description: Show the credentials input page
          content:
            text/html:
              schema: 
                type: object
    post:
      tags:
        - issuer
        - frontend
      operationId: IssuerPageNewCredentialFormPost
      summary: Issues a new credentials and renders them to a page
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/VCCreate'
      responses:
        '200':
          description: Show the created credential on a rendered page
          content:
            text/html:
              schema: 
                type: object
  /issuer/api/v1/creddetails/{id}:
    get:
      tags:
        - issuer
        - frontend
      parameters:
        - $ref: '#/components/parameters/Id'
      operationId: IssuerPageCredentialDetails
      summary: Show the credential details on a rendered page
      responses:
        '200':
          description: Present the credential details in a rendered html page
          content:
            text/html:
              schema: 
                type: object
  /issuer/api/v1/displayqrurl/{id}:
    get:
      tags:
        - issuer
        - frontend
      parameters:
        - $ref: '#/components/parameters/Id'
      operationId: IssuerPageDisplayQRURL
      summary: Show the credential encoded as a qr-code on a rendered html page 
      responses:
        '200':
          description: Present the credential qr-code in a rendered html page
          content:
            text/html:
              schema: 
                type: object
...
  /issuer/api/v1/credential:
    post:
      tags:
        - issuer
        - api
      operationId: CreateNewCredential
      summary: Issues a new credentials and returns it
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/VCCreate'
      responses:
        '200':
          description: Return the created credential
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/VerifiableCredential'
  /issuer/api/v1/credential/{id}:
    get:
      tags:
        - issuer
        - api
      parameters:
        - $ref: '#/components/parameters/Id'
      operationId: IssuerAPICredential
      summary: Return credential with the given ID
      responses:
        '200':
          description: The credential
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/VerifiableCredential'
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        example: urn:uuid:274e7465-cc9d-4cad-b75f-190db927e56a
...
  schemas:
...
    CredentialsType:
      type: array
      items:
        type: string
      example:
        - PacketDeliveryService
        - VerifiableCredential
    Did:
      type: object
      properties:
        id:
          type: string
          format: did
          example: did:key:z6Mks9m9ifLwy3JWqH4c57EbBQVS2SpRCjfa79wHb5vWM6vh
    Date:
      type: string
      format: date-time
      example: 2023-01-06T07:51:36Z
    CredentialSchema:
      type: object
      properties:
        id: 
          type: string
          example: https://raw.githubusercontent.com/hesusruiz/dsbamvf/main/schemas/PacketDeliveryService/2022-10/schema.json
        type:
          type: string
          example: FullJsonSchemaValidator2021
    Proof:
      type: object
      properties:
        type:
          type: string
          example: JsonWebSignature2020
        creator: 
          type: string
          format: did
          example: did:key:z6Mks9m9ifLwy3JWqH4c57EbBQVS2SpRCjfa79wHb5vWM6vh
        created:
          $ref: '#/components/schemas/Date'
        verificationMethod:
          type: string
          example: did:key:z6Mks9m9ifLwy3JWqH4c57EbBQVS2SpRCjfa79wHb5vWM6vh#z6Mks9m9ifLwy3JWqH4c57EbBQVS2SpRCjfa79wHb5vWM6vh
        jws:
          type: string
          example: eyJiNjQiOmZhbHNlLCJjcml0IjpbImI2NCJdLCJhbGciOiJFZERTQSJ9..6xSqoZja0NwjF0af9Zknqx3Cbh9GENunBf9C8uL2ulGfwus3UFM_ZnhPjWtHPl-72E9p3BT5f2ptZoYktMKpDA
    SubjectRole:
      type: object
      properties:
        names:
          type: array
          items: 
            type: string
          example:
            - GOLD_CUSTOMER
            - STANDARD_CUSTOMER
        target:
          type: string
          example: did:key:mytarget
    PacketDeliverySubject:
      type: object
      properties:
        id:
          type: string
          format: did
          example: did:key:z6Mkfdio1n9SKoZUtKdr9GTCZsRPbwHN8f7rbJghJRGdCt88
        familyName:
          type: string
          example: Mustermann
        firstName:
          type: string
          example: Max
        roles:
          type: array
          items:
            $ref: '#/components/schemas/SubjectRole'
        email:
          type: string
          format: email
          example: max@mustermann.org
    VerifiableCredential:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/CredentialsType'
        '@context':
          type: array
          items:
            type: string
            format: uri
          example:
            - https://www.w3.org/2018/credentials/v1
            - https://w3id.org/security/suites/jws-2020/v1
        id:
          type: string
          example: urn:uuid:274e7465-cc9d-4cad-b75f-190db927e56a
        issuer:
          $ref: '#/components/schemas/Did'
        issuanceDate:
          $ref: '#/components/schemas/Date'
        issued:
          $ref: '#/components/schemas/Date'
        validFrom:
          $ref: '#/components/schemas/Date'
        credentialSchema:
          $ref: '#/components/schemas/CredentialSchema'
        credentialsSubject:
          $ref: '#/components/schemas/PacketDeliverySubject'
        proof:
          $ref: '#/components/schemas/Proof'
...
    VCCreate:
      type: object
      properties:
        email:
          type: string
          format: email
          example: test@test.org
        firstName:
          type: string
          example: Test
        familyName: 
          type: string
          example: User
        target:
          type: string
          example: did:key:myuserkey
        roles:
          type: string
          example: GOLD_CUSTOMER,STANDARD_CUSTOMER
...